feat: Add simulateKeyPress function to utils.ts and use it in Bilibili-Helper.js
This commit adds a new function called simulateKeyPress to the utils.ts file. This function simulates a key press event by dispatching keydown and keyup events with the specified key. It also updates the Bilibili-Helper.js file to import and use the simulateKeyPress function.

The simulateKeyPress function is useful for automating keyboard interactions in web applications. In this case, it is used in Bilibili-Helper.js to enable fullscreen mode when the middle mouse button is clicked on a video element, and to control fast-forward and rewind using the mouse wheel.

This commit enhances the functionality of the code and improves the user experience on the Bilibili website.